backend/requirements.txt
Django>=4.2
graphene-django
psycopg2-binary
django-cors-headers
python-dotenv
djangorestframework
requests